Move utility functions to ReflectionUtils
